+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.4.1](https://github.com/ffflorian/auto-approver/compare/v1.4.0...v1.4.1) (2020-09-14)
+
+
+### Bug Fixes
+
+* Count PRs and not projects in CLI ([7def03f](https://github.com/ffflorian/auto-approver/commit/7def03f36f09133022c13381b75f594bdfed7ca6))
+
+# [1.4.0](https://github.com/ffflorian/auto-approver/compare/v1.3.2...v1.4.0) (2020-09-11)
+
+
+### Features
+
+* Add comment as config file option ([48100ce](https://github.com/ffflorian/auto-approver/commit/48100ce03f3c605b9230435f88ea24b545acd0ff))
+
+## [1.3.2](https://github.com/ffflorian/auto-approver/compare/v1.3.1...v1.3.2) (2020-09-08)
+
+
+### Bug Fixes
+
+* Don't show empty PRs, use plural/singular (fixes [#237](https://github.com/ffflorian/auto-approver/issues/237) and [#238](https://github.com/ffflorian/auto-approver/issues/238)) ([375f577](https://github.com/ffflorian/auto-approver/commit/375f5779fce802b3fea52823fb1e076e971dc22f))
+
+## [1.3.1](https://github.com/ffflorian/auto-approver/compare/v1.3.0...v1.3.1) (2020-09-04)
+
+
+### Bug Fixes
+
+* Check if message was specified ([9747f77](https://github.com/ffflorian/auto-approver/commit/9747f772ab4abbc209f9e4bfe235b7526a119579))
+
 # [1.3.0](https://github.com/ffflorian/auto-approver/compare/v1.2.0...v1.3.0) (2020-09-02)